refactor(ui-editor): reorganize tags/widgets/ subdirectory (anchor.ts, placeholder.ts)
