+++ packages/ui/ui-editor/src/extensions/tags/widgets/index.ts
@@ -1,0 +1,6 @@
+//
+// Copyright 2025 DXOS.org
+//
+
+export * from './anchor';
+export * from './placeholder';
